헤더부분 분리
