+++ p-1/frontend/index.html
@@ -8,6 +8,6 @@
     <script defer type="module" src="./static/js/index.js"></script>
 </head>
 <body>
-      <div id="root"></div>
+    <div id="root"></div>
 </body>
 </html>
