--- p-1/frontend/static/js/pages/menu.js
@@ -1,10 +1,0 @@
-export default class {
-    constructor() {
-        document.title = "Menu";
-    }
-    async getHtml() {
-        return `
-            <h1>This is Menu Page</h1>
-        `;
-    }
-}
